+++ src/schematics/rucken-app/index.ts
@@ -131,7 +131,7 @@ function templateSources(options: NormalizedSchema) {
 function updateScriptsPostinstallSh(options: NormalizedSchema): Rule {
   return updateFileInTree(join(normalize(options.workspace), 'scripts', 'postinstall.sh'), data => {
     const content = readFileSync(
-      `./files/rucken/todo/scripts/postinstall.sh`.replace('{localPath}', options.basePath)
+      `${__dirname}/../../../files/rucken/todo/scripts/postinstall.sh`.replace('{localPath}', options.basePath)
     ).toString();
     if (clearText(data).indexOf(clearText(content)) === -1) {
       data = data + '\n' + content;
@@ -142,7 +142,7 @@ function updateScriptsPostinstallSh(options: NormalizedSchema): Rule {
 function updateScriptsPatchJs(options: NormalizedSchema): Rule {
   return updateFileInTree(join(normalize(options.workspace), 'scripts', 'patch.js'), data => {
     const original = readFileSync(
-      `./files/rucken/todo/scripts/patch.js`.replace('{localPath}', options.basePath)
+      `${__dirname}/../../../files/rucken/todo/scripts/patch.js`.replace('{localPath}', options.basePath)
     ).toString();
     const content = '(function(){\n' + original + ';\n})();';
     if (original && clearText(data).indexOf(clearText(content)) === -1) {
@@ -155,7 +155,7 @@ function updatePackageJson(tree: Tree, options: NormalizedSchema): Rule {
   return updateJsonInTree(join(normalize(options.workspace), 'package.json'), packageJson => {
     const templatePackageJson = JSON.parse(
       readFileSync(
-        `./files/rucken/todo/package.json`.replace('{localPath}', options.basePath)
+        `${__dirname}/../../../files/rucken/todo/package.json`.replace('{localPath}', options.basePath)
       ).toString()
     );
     packageJson.author = options.fullAuthorObject;